requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
opensimplex>=0.4
python-pptx>=0.6.21
